Avoid condition in amd64 popcount.
This makes it not possible to inline, but does show that it benches
faster:

BenchmarkASM-4             2.29          2.64          +15.28%
BenchmarkASMIndirect-4     4.23          3.04          -28.13%
BenchmarkGeneric-4         3.25          3.18          -2.15%

The numbers are small, so things fluctuate a bit (as you can see in the
results for asm and generic on this 15s bench), but most importantly,
the asm version with conditional was typically slower than the pure go
version.  This new variant is still slower than calling the asm version
directly, but is at least typically faster than the generic version.